requirements.txt
requests
python-nmap
colorama